Make sample_data fixture available
